Project creation and Step 1 completed
Roll-a-Ball project created and Step 1 of the tutorial finished. This
included saving the scene "MiniGame" and putting it into the "_Scenes"
folder, making a plane and naming it "Ground", making a sphere and
naming it "Player", and making a material and naming it "Background"
before putting it into the "Materials" Folder. The "Background" material
was then colored a dark blue and put on the "Ground" plane. The
directional light's y-axis rotation was changed to 60.